requirements.txt
python-dotenv~=1.0.1
colorama~=0.4.6
requests~=2.32.3